pyproject.toml
[project]
name = "fastflow"
version = "0.2.0"
description = "A Python library wrapping AntV X6 for FastHTML - build visual workflow editors with Python using fastcore patterns"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Sebastian Gaseretto"}
]
keywords = ["fasthtml", "antv-x6", "workflow", "node-editor", "visual-programming", "htmx", "fastcore", "callbacks"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "python-fasthtml>=0.12.0",
    "fastcore>=1.5.0",
    "plum-dispatch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/sgaseretto/fastflow"
Repository = "https://github.com/sgaseretto/fastflow"
Documentation = "https://github.com/sgaseretto/fastflow#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastflow"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
